frontend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.22
asyncpg==0.28.0
psycopg2-binary==2.9.9
alembic==1.12.0

# Data Processing
pandas==2.1.1
numpy==1.26.0
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.0

# Development
pytest==7.4.2
black==23.9.1
mypy==1.6.1